requirements.txt
streamlit
pymilvus
torch
openai-clip
Pillow